# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Exetel Laravel Rest API Assingment  Documentation",
        "description": "Use this as x-api-key `GzK2!H4m@*Me%a-zPL4TYz#JBw26MsMMjJCSbr+=BVqp&3f8gyRk%TF2gaVkN8%$ydR$3rE#F*Eue*5Fj4GdD9he335?R+zpDvQR7m6v84YMBe6#38SDU8vvfvnWfxUx6+=jdhMtwfv4^7RbuhuF+B_$&M5ztsswcAMQ^%Fd4Wc*UD@#AtcxYF@Jey&H3M_JW23s4CYnk#?yrtZM_t!vdd2Z7L8pYat%Y2vMCPv#K&uQEZBbnCL3!6c579f%C%*Q` ",
        "contact": {
            "email": "dinethwa@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/user/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login User to System",
                "description": "Login user and get access token",
                "operationId": "userLogin",
                "parameters": [
                    {
                        "parameter": "x-api-key",
                        "name": "x-api-key",
                        "in": "header",
                        "description": "Application Verification Token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data packet for User Login",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            },
                            "example": {
                                "username": "abcd",
                                "password": "123456"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Welcome"
                                        },
                                        "data": {
                                            "properties": {
                                                "access_token": {
                                                    "type": "string",
                                                    "example": "sample-token"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid Credentials"
                                        },
                                        "userMessage": {
                                            "type": "string",
                                            "example": "user Message"
                                        },
                                        "errorCode": {
                                            "type": "string",
                                            "example": "error Code"
                                        },
                                        "metaData": {
                                            "type": "object",
                                            "example": "error Code"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}